requirements.txt
tensorflow==2.15.0
tensorflow-privacy==0.9.0
flwr==1.7.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
shap==0.44.0
matplotlib==3.7.2
pyyaml==6.0.1